pipeline-manager: break up api.rs into per-api-object source files
